Add g:CommandTMaxCachedDirectories option
This allows the user to specify an upper limit on the number of caches
which Command-T will use when changing directories. The default of 1
preserves the existing behavior, meaning that each time you change into
a different directory, Command-T will scan it.

Set to a higher number to increase the number of cached directories, or
to 0 for no limit at all.

Signed-off-by: Wincent Colaiuta <win@wincent.com>